Hookify the code editor code
