+++ src/index.ts
@@ -1,8 +1,7 @@
 import { addFilter } from '@wordpress/hooks';
 import { registerPlugin } from '@wordpress/plugins';
 
-import { setupViewOnlyMode } from './components/view-only-blocks-mode';
-import { ReadOnlyCodeEditor } from '@/components/read-only-code-editor';
+import { setupViewOnlyMode } from './components/view-only-mode';
 import { RTCSettingsPanel } from '@/components/rtc-settings-panel';
 import { WEBSOCKET_URL } from '@/utilities/config';
 import { Logger } from '@/utilities/logger';
@@ -26,7 +25,3 @@ registerPlugin( 'vip-rtc-settings-panel', {
 } );
 
 setupViewOnlyMode();
-
-registerPlugin( 'vip-rtc-read-only-code-editor', {
-	render: ReadOnlyCodeEditor,
-} );
